Demo/requirements.txt
streamlit==1.28.0
pandas==2.0.3
pdfplumber==0.10.3
matplotlib==3.7.2
openpyxl==3.1.2
reportlab==4.0.6
Pillow==10.0.1